+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+
+## [0.15.0] - 2023-09-12
 ### Added
 - Add loadEnvrc command which lets the user load any .envrc file
 
@@ -121,7 +123,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.1] - 2021-11-23
 - Initial release
 
-[Unreleased]: https://github.com/direnv/direnv-vscode/compare/v0.14.0...HEAD
+[Unreleased]: https://github.com/direnv/direnv-vscode/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/direnv/direnv-vscode/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/direnv/direnv-vscode/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/direnv/direnv-vscode/compare/v0.12.2...v0.13.0
 [0.12.2]: https://github.com/direnv/direnv-vscode/compare/v0.12.1...v0.12.2